+++ src/raw_bmp.rs
@@ -1,18 +1,24 @@
-use embedded_graphics::{prelude::*, primitives::Rectangle};
-
 use crate::{
     color_table::ColorTable,
     header::{Bpp, Header},
-    pixels::Pixels,
-    raw_pixels::RawPixels,
-    ParseError, RawPixel,
+    raw_iter::RawPixels,
+    ChannelMasks, ParseError,
 };
 
-/// A BMP-format bitmap.
+/// Low-level access to BMP image data.
+///
+/// This struct can be used to access the image data in a BMP file at a lower level than with the
+/// [`Bmp`](crate::Bmp) struct. It doesn't do automatic color conversion and doesn't apply the color
+/// table, if it is present in the BMP file. For images with a color table the iterator returned by
+/// [`pixels`](Self::pixels) will instead return the color indices, that can be looked up manually
+/// using the [`ColorTable`] struct.
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
 pub struct RawBmp<'a> {
     /// Image header.
     header: Header,
+
+    /// Color type.
+    pub(crate) color_type: ColorType,
 
     /// Color table for color mapped images.
     color_table: Option<ColorTable<'a>>,
@@ -26,97 +32,93 @@ impl<'a> RawBmp<'a> {
     ///
     /// The created object keeps a shared reference to the input and does not dynamically allocate
     /// memory.
-    ///
-    /// In contrast to the [`from_slice`] constructor no color type needs to be specified when
-    /// calling this method. This will disable all functions that requires a specified color type,
-    /// like the [`pixels`] method.
-    ///
-    /// [`from_slice`]: #method.from_slice
-    /// [`pixels`]: #method.pixels
     pub fn from_slice(bytes: &'a [u8]) -> Result<Self, ParseError> {
         let (_remaining, (header, color_table)) = Header::parse(bytes)?;
 
+        let color_type = ColorType::from_header(&header)?;
+
+        let data_length = header.bytes_per_row() * header.image_size.height as usize;
+
         let image_data = &bytes
-            .get(header.image_data_start..)
+            .get(header.image_data_start..header.image_data_start + data_length)
             .ok_or(ParseError::UnexpectedEndOfFile)?;
 
         Ok(Self {
             header,
+            color_type,
             color_table,
             image_data,
         })
     }
 
-    /// Returns the size of this image in pixels.
-    pub fn size(&self) -> Size {
-        self.header.image_size
-    }
-
-    /// Returns the BPP (bits per pixel) for this image.
-    pub fn color_bpp(&self) -> Bpp {
-        self.header.bpp
-    }
-
-    /// Gets the color table associated with the image.
-    pub(crate) fn color_table(&self) -> Option<&ColorTable<'a>> {
+    /// Returns the color table associated with the image.
+    pub const fn color_table(&self) -> Option<&ColorTable<'a>> {
         self.color_table.as_ref()
     }
 
     /// Returns a slice containing the raw image data.
-    pub fn image_data(&self) -> &'a [u8] {
+    pub const fn image_data(&self) -> &'a [u8] {
         self.image_data
     }
 
     /// Returns a reference to the BMP header.
-    pub fn header(&self) -> &Header {
+    pub const fn header(&self) -> &Header {
         &self.header
     }
 
     /// Returns an iterator over the raw pixels in the image.
     ///
-    /// The iterator returns the raw pixel colors as `u32` values. To automatically convert the raw
-    /// values into the color specified by `C` use [`pixels`] instead.
-    ///
-    /// [`pixels`]: #method.pixels
-    pub fn pixels<'b>(&'b self) -> RawPixels<'b, 'a> {
+    /// The iterator returns the raw pixel colors as [`u32`] values.  To automatically convert the
+    /// raw values into [`embedded_graphics`] color types use [`Bmp::pixels`](crate::Bmp::pixels)
+    /// instead.
+    pub fn pixels(&self) -> RawPixels<'_> {
         RawPixels::new(self)
     }
+}
 
-    /// Returns the row length in bytes.
-    ///
-    /// Each row in a BMP file is a multiple of 4 bytes long.
-    pub(crate) fn bytes_per_row(&self) -> usize {
-        let bits_per_row =
-            self.header.image_size.width as usize * usize::from(self.header.bpp.bits());
+#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+pub enum ColorType {
+    Index1,
+    Index4,
+    Index8,
+    Rgb555,
+    Rgb565,
+    Rgb888,
+    Xrgb8888,
+}
 
-        (bits_per_row + 31) / 32 * (32 / 8)
-    }
-
-    pub(crate) fn draw<D>(&self, target: &mut D) -> Result<(), D::Error>
-    where
-        D: DrawTarget,
-        D::Color: From<<D::Color as PixelColor>::Raw>,
-    {
-        if self.color_bpp().bits() <= 8 {
-            if let Some(color_table) = self.color_table {
-                target.fill_contiguous(
-                    &Rectangle::new(Point::zero(), self.size()),
-                    self.pixels().map(|RawPixel { color, .. }| {
-                        color_table
-                            .get_raw::<<<D as DrawTarget>::Color as PixelColor>::Raw>(color)
-                            .unwrap_or_else(|| RawData::from_u32(0)) //TODO: how should invalid color indices be handled
-                            .into()
-                    }),
-                )
-            } else {
-                // Don't try to draw anything if the color table is missing.
-                Ok(())
+impl ColorType {
+    pub(crate) fn from_header(header: &Header) -> Result<ColorType, ParseError> {
+        Ok(match header.bpp {
+            Bpp::Bits1 => ColorType::Index1,
+            Bpp::Bits4 => ColorType::Index4,
+            Bpp::Bits8 => ColorType::Index8,
+            Bpp::Bits16 => {
+                if let Some(masks) = header.channel_masks {
+                    match masks {
+                        ChannelMasks::RGB555 => ColorType::Rgb555,
+                        ChannelMasks::RGB565 => ColorType::Rgb565,
+                        _ => return Err(ParseError::UnsupportedChannelMasks),
+                    }
+                } else {
+                    // According to the GDI docs the default 16 bpp color format is Rgb555 if no
+                    // color masks are defined:
+                    // https://docs.microsoft.com/en-us/windows/win32/api/wingdi/ns-wingdi-bitmapinfoheader
+                    ColorType::Rgb555
+                }
             }
-        } else {
-            target.fill_contiguous(
-                &Rectangle::new(Point::zero(), self.size()),
-                Pixels::new(self.pixels()).map(|Pixel(_, color)| color),
-            )
-        }
+            Bpp::Bits24 => ColorType::Rgb888,
+            Bpp::Bits32 => {
+                if let Some(masks) = header.channel_masks {
+                    if masks == ChannelMasks::RGB888 {
+                        ColorType::Xrgb8888
+                    } else {
+                        return Err(ParseError::UnsupportedChannelMasks);
+                    }
+                } else {
+                    ColorType::Xrgb8888
+                }
+            }
+        })
     }
 }