+++ tests/chessboard-8px-1bit.rs
@@ -22,6 +22,12 @@ fn chessboard_8px_1bit() {
         }
     );
 
+    let color_table = bmp.color_table().unwrap();
+    assert_eq!(color_table.len(), 2);
+    assert_eq!(color_table.get(0), Some(Rgb888::BLACK));
+    assert_eq!(color_table.get(1), Some(Rgb888::WHITE));
+    assert_eq!(color_table.get(2), None);
+
     assert_eq!(bmp.image_data().len(), 94 - 62);
 }
 
@@ -209,4 +215,18 @@ fn chessboard_8px_1bit_0colors() {
     ];
 
     assert_eq!(pixels, expected);
+}
+
+#[test]
+// A regression was found from the original fix, described in <https://github.com/embedded-graphics/tinybmp/issues/18#issuecomment-1101600500>
+fn issue_18() {
+    let bmp = Bmp::<'_, BinaryColor>::from_slice(include_bytes!(
+        "../tests/chessboard-8px-1bit-0colors.bmp"
+    ))
+    .expect("Failed to parse");
+
+    let color_table = bmp.as_raw().color_table().unwrap();
+    assert_eq!(color_table.len(), 2);
+    assert_eq!(color_table.get(0), Some(Rgb888::BLACK));
+    assert_eq!(color_table.get(1), Some(Rgb888::WHITE));
 }