+++ tests/colors_8bpp_indexed.rs
@@ -6,6 +6,11 @@ fn colors_8bpp_indexed() {
     let bmp = Bmp::<'_, Rgb888>::from_slice(include_bytes!("./colors_8bpp_indexed.bmp"))
         .expect("Failed to parse");
 
+    assert!(
+        bmp.as_raw().color_table().is_some(),
+        "there should be a color table for this image"
+    );
+
     let pixels: Vec<u32> = bmp
         .pixels()
         .map(|Pixel(_pos, color)| color.into_storage())